rename files and fix a bug in expansion
